Generalize indexed
