+++ fast_llm/data/dataset/gpt/indexed.py
@@ -1,0 +1,60 @@
+import abc
+import typing
+
+import numpy as np
+
+from fast_llm.data.data.gpt.data import GPTData
+from fast_llm.data.dataset.gpt.config import GPTSamplingConfig
+from fast_llm.data.dataset.indexed import ConcatenatedIndexedDataset, IndexedDataset, IndexedDatasetSlice
+
+if typing.TYPE_CHECKING:
+    from fast_llm.data.dataset.gpt.sampled import GPTSampledIndexedDataset
+
+
+class GPTIndexedDataset(IndexedDataset):
+    """
+    A GPT dataset containing a list of samples.
+    """
+
+    # def get(self, index: int, offset: int = 0, length: int | None = None):
+    #    pass
+
+    # def __len__(self) -> int:
+    #    """
+    #    Number of documents in the dataset.
+    #    Can be calculated from document sizes but may be overridden if there is a better method.
+    #    """
+    #    return len(self.get_document_sizes())
+
+    @abc.abstractmethod
+    def get_document_sizes(self) -> np.ndarray:
+        """
+        The size of each document in the dataset.
+        The resulting array could be very large, so this method should be called cautiously,
+        and derived classes should try to avoid holding the whole array im memory.
+        """
+
+    def sample(self, config: GPTSamplingConfig, data: GPTData) -> "GPTSampledIndexedDataset":
+        from fast_llm.data.dataset.gpt.sampled import GPTSampledIndexedDataset
+
+        return GPTSampledIndexedDataset(self, config, data)
+
+
+class GPTDatasetSlice(IndexedDatasetSlice, GPTIndexedDataset):
+    """
+    A GPT dataset, which reads samples from (a split of) a `MMapIndexedDataset` pointing to a GPT dataset.
+    """
+
+    _dataset: GPTIndexedDataset
+
+    def get_document_sizes(self):
+        # TODO: This can be really big.
+        return self._dataset.get_document_sizes()[self._begin : self._end]
+
+
+class GPTConcatenatedDataset(ConcatenatedIndexedDataset, GPTIndexedDataset):
+    _datasets: list[GPTIndexedDataset]
+
+    def get_document_sizes(self) -> np.ndarray:
+        # TODO: This can be really big.
+        return np.concatenate([dataset.get_document_sizes() for dataset in self._datasets])
